move bytecodes to static.ts
